bin: add module rc.js
* Move shared code into internal module `./rc.js`.

* This internal module will be used by executables `ijs`, `ijsinstall`,
  `ijsnotebook` and `ijsconsole`.
